requirements.txt
llama-cpp-python
cached_property
farm-haystack
fastapi
streamlit
uvicorn
farm-haystack[weaviate]
pypdf
sentence_transformers
haystack-ai
torch
python-dotenv